Updated script to use regex
The script now uses a regex pattern to match the lines, filtering out
any public, static, and/or final fields, as we don't want to turn these
into properties.  Also handles default values on the fields (i.e. the
field being declared with an assignment as well).  Also changes the
format of the template slightly, and groups the getters separately from
the setters.